split up import-export file
